requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bleach==4.1.0
Django==4.0.3
django-bootstrap4==22.1
django-filter==21.1
django-summernote==0.8.20.0
django-taggit==2.1.0
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
webencodings==0.5.1
